chore: release v3.2.1
Consolidated maintenance release:

- Enforce a 300-LOC cap via the max-lines oxlint rule (error) and
  document it in CONTRIBUTING under Code Quality.
- Require type aliases over interface (consistent-type-definitions: type).
- Split deskew.ts -> deskew-angles.ts and canvas-processor.ts ->
  canvas-regions.ts + canvas-io.ts to fit the cap.
- Modularize comparison and canvas-processor test suites into subfolders.
- Fix pre-commit so formatting is scoped to staged files (no repo-wide
  reformat sweeping unrelated files into a commit).
- Fix the vanilla-HTML demo + README to load opencv.js before initRuntime.
