+++ jsr.json
@@ -1,6 +1,6 @@
 {
   "name": "@snowfluke/ppu-ocv",
-  "version": "3.2.0",
+  "version": "3.2.1",
   "license": "MIT",
   "exports": {
     ".": "./src/index.ts",
