+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ppu-ocv",
-  "version": "3.2.0",
+  "version": "3.2.1",
   "description": "A type-safe, modular, chainable image processing library built on top of OpenCV.js with a fluent API leveraging pipeline processing.",
   "keywords": [
     "open-cv",
